readerController load audio into map<blockID, audioData>
The change is needed as MSVC doesn't allow for a vector of a map
of a non-copyable object.

We can't use a list, as apparently several calls to loadAudio need
to write the blocks of different audio groups into the same position
in the output vector. std::list doesn't have random access.

Using a map now, so we can keep the previous behavior. The int64_t key
is only used in construction, and can be disregarded by the loading
code.