restore symlink fixing actions for Mac
In a refactoring of the DELAY_LOAD stuff of Windows it seems the
Mac version was broken. On Mac there is a link to libvideostitch,
which is just a symlink to either _cuda or _opencl versions.

If the the symlink is currently pointing at the wrong one (i.e.
not the one matching our GPU), the symlink is changed and the app
is quit, user asked to launch again.

There seems to have been a refactoring done on Windows which de-
activated DELAY_LOAD completely on Mac, which made some of the
symlink changing code not compile.

Adding a new C Macro for Mac symlinks and allowing for USE_DELAY_LOAD
on Mac as well, in CMake, to restore previous behavior.